requirements.txt
google-generativeai
google-api-python-client
python-dotenv
requests
beautifulsoup4
pypdf
flask
markdown